Code Tidy: Create base policy to remove duplication
