+++ app/policies/integral/user_policy.rb
@@ -1,17 +1,6 @@
 module Integral
   # Handles User authorization definitions
-  class UserPolicy
-    attr_reader :user, :instance
-
-    # Initializes a UserPolicy instance which is used to handle authorization for User
-    #
-    # @param user [User] user policy is authorizing against
-    # @param instance [User] instance policy is authorizing against
-    def initialize(user, instance)
-      @user = user
-      @instance = instance
-    end
-
+  class UserPolicy < BasePolicy
     # @return [Boolean] if user is allowed to perform an update
     def update?
       user.has_role?(:user_manager) || instance == user
@@ -21,13 +10,5 @@ module Integral
     def manager?
       user.has_role?(:user_manager)
     end
-
-    alias_method :create?, :manager?
-    alias_method :new?, :manager?
-    alias_method :destroy?, :manager?
-    alias_method :index?, :manager?
-
-    alias_method :edit?, :update?
-    alias_method :show?, :update?
   end
 end
